edit min hash
