+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-search",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "This package uses seraching algorithm to find similar strings",
   "main": "index.js",
   "scripts": {
